Move abandon_io_channel to test folder
