+++ test/shared/emp-tool-test.h
@@ -1,2 +1,4 @@
-#include "test/shared/plain_circ.h"
-#include "test/shared/plain_prot.h"
+#include "test/shared/execution/plain_circ.h"
+#include "test/shared/execution/plain_prot.h"
+
+#include "test/shared/io/abandon_io_channel.h"
