+++ test/shared/io/abandon_io_channel.h
